refactor: rename defaultConstructGetUri to url
